getBalance refactor, but balance mechanism still not implemented

+++ internal/model/model.go
@@ -1,7 +1,7 @@
 package model
 
 type Balance struct {
-	Gold   uint16
+	Gold   uint
 	Silver uint8
 	Bronze uint8
 }
