+++ internal/repo/model.go
@@ -1,5 +1,5 @@
 package repo
 
 type coins struct {
-	balance uint16 `db:"balance"`
+	balance uint `db:"balance"`
 }
